requirements.txt
pytorch-transformers
numpy
lmdb==0.94
tqdm==4.31.1
easydict==1.9
PyYAML==5.1.2
jsonlines==1.2.0
json-lines==0.5.0
matplotlib
Cython
python-prctl
msgpack
msgpack-numpy
opencv-python==4.2.0.34
scikit-image
yacs
tensorflow
tensorboardX 
tensorflow 
tensorpack

-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.7.1+cu110 
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.2+cu110